+++ src/pages/onboarding/OnboardingLifestylePage.tsx
